Move :.: into its own module.
